Refactor Marquee component: move to magicui directory and update imports; remove old marquee implementation; enhance Tailwind CSS animations
